change fish-detection heuristic
- slightly adjust the X axis range considered for tail detection
- throw out hacked up anti-vignette correction
- replace vertical percentile threshold with threshold from local maxima
- do more work in uint8 space rather than float32 for speed

This still considers the overall detection problem independently for
each pixel column, i.e. position in Y axis.

The new effect is a bit like an edge detection or high-pass filter to
find smaller dark areas relative to their neighborhood. With a narrow
kernel width it seems to work well enough (and fast) on the tail and
is benign on the anterior of the fish.