make movie analysis tolerate apparent 4th light level
- don't assert len(light_bases) <= 3
- clamp light_index to max 2 when converting to off/CS/US labels

this helps if the global illumination drifts a bit and the max US
level is brighter late in the movie.

if the brightness drifts so much that the other levels overlap,
the current method will still fail.  that would require some other
self-tuning to adjust the thresholds used for each intensity plateau.